# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Biomark AI API
  version: 1.0.0
  description: API pública de salud preventiva. La IA no diagnostica ni prescribe.
servers:
  - url: https://api.example.com
    description: VPS de producción
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /api/chat:
    post:
      tags: [Chat]
      summary: Envía una consulta clínica preventiva
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ChatRequest' } } } }
      responses: { '200': { description: Respuesta de Biomark AI } }
  /api/voice:
    post:
      tags: [Chat]
      summary: Envía audio y recibe transcripción/respuesta
      requestBody: { required: true, content: { multipart/form-data: { schema: { type: object, required: [archivo], properties: { archivo: { type: string, format: binary }, session_id: { type: string, format: uuid } } } } } }
      responses: { '200': { description: Respuesta de voz } }
...
  /api/vision:
    post:
      tags: [Salud]
      summary: Analiza una imagen de piel o garganta
      parameters: [{ name: tipo, in: query, required: true, schema: { type: string, enum: [piel, garganta] } }]
      requestBody: { required: true, content: { multipart/form-data: { schema: { type: object, required: [archivo], properties: { archivo: { type: string, format: binary } } } } } }
      responses: { '200': { description: Resultado de apoyo visual } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  schemas:
...
    ChatRequest: { type: object, required: [message], properties: { message: { type: string }, session_id: { type: string, format: uuid }, latitude: { type: number }, longitude: { type: number } } }
...